requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.44.2
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
Eel==0.17.0
future==1.0.0
gevent==24.10.3
gevent-websocket==0.10.1
greenlet==3.1.1
idna==3.10
numpy==2.1.1
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
pycparser==2.22
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.2.0
pywin32-ctypes==0.2.3
requests==2.32.3
scipy==1.14.1
setuptools==73.0.1
sounddevice==0.5.0
ttkbootstrap==1.10.1
urllib3==2.2.3
whichcraft==0.6.1
zope.event==5.0
zope.interface==7.1.0
